fix: data_mapping.py works separetly
