+++ .gitignore
@@ -214,3 +214,21 @@ vignettes/*.pdf
 
 # R Environment Variables
 .Renviron
+
+#===============================================================================================
+# Project specific
+#===============================================================================================
+# Avast temporary files
+_avast_/
+
+# Additional data and results
+*.csv
+*.xlsx
+*.xls
+*.sav
+*.rds
+*.pkl
+
+# Temporary files
+*.tmp
+*.bak
